- added .gitignore file (Created by https://www.gitignore.io) - bonigarcia updated to version 3.6.2 - broken imports was removed - change configuration of maven-compiler-plugin, configuration parameter "suiteXmlFile" was change to curent test file name - MainPage.java submitSearch() method returns correct page - chenge type void methods in the ResultPage - fix double cast in ExtentTestManager page (((int) (long) (Thread.currentThread().getId()))) - BasePage moved to base/pages packedge - TestListener simplified check driver exist, unused methods was deleted - Removed duplicates variable from systemPropertyVariables - Move business logic from test methods to Page objects - Change AfterMethod logic (if page title didn't contain domain name from properties, we should go back to main page)

+++ src/main/java/base/utils/listeners/ExtentTestManager.java
@@ -11,11 +11,11 @@ public class ExtentTestManager {
     static ExtentReports extent = ExtentManager.getReporter();
 
     public static synchronized ExtentTest getTest() {
-        return (ExtentTest) extentTestMap.get((int) (long) (Thread.currentThread().getId()));
+        return (ExtentTest) extentTestMap.get((int) (Thread.currentThread().getId()));
     }
 
     static synchronized void startTest(String testName, String desc) {
         ExtentTest test = extent.startTest(testName, desc);
-        extentTestMap.put((int) (long) (Thread.currentThread().getId()), test);
+        extentTestMap.put((int) (Thread.currentThread().getId()), test);
     }
 }
